switched to size_t

+++ lib/transfer.cpp
@@ -254,7 +254,7 @@ namespace quda {
     ColorSpinorField &coarse(*coarse_tmp_h);
 
     // compute the coarse grid point for every site (assuming parity ordering currently)
-    for (unsigned int i = 0; i < fine.Volume(); i++) {
+    for (size_t i = 0; i < fine.Volume(); i++) {
       // compute the lattice-site index for this offset index
       fine.LatticeIndex(x, i);
